+++ test/main.cpp
@@ -3,12 +3,12 @@
 
 using namespace Gularen;
 
-void print(Node* node, unsigned int depth = 0) {
-	for (unsigned int i = 0; i < depth; i += 1) {
+void print(Node* node, size_t depth = 0) {
+	for (size_t i = 0; i < depth; i += 1) {
 		std::cout << "  ";
 	}
 	node->print();
-	for (unsigned int i = 0; i < node->children.size(); i += 1) {
+	for (size_t i = 0; i < node->children.size(); i += 1) {
 		print(node->children[i], depth + 1);
 	}
 }
